// src/api/generated/models/SupplierUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Update supplier
 */
export type SupplierUpdate = {
    name?: (string | null);
    contact_name?: (string | null);
    email?: (string | null);
    phone?: (string | null);
    address?: (string | null);
    website?: (string | null);
    payment_terms?: (string | null);
    notes?: (string | null);
    is_active?: (boolean | null);
};